+++ packages/frontend/src/components/Explorer/SaveChartButton/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState } from 'react';
 import { useAddVersionMutation } from '../../../hooks/useSavedQuery';
 import { useExplorerContext } from '../../../providers/ExplorerProvider';
-import ChartCreateModal from '../../SavedQueries/ChartCreateModal';
+import ChartCreateModal from '../../common/modal/ChartCreateModal';
 import { SaveButton } from './SaveChartButton.styles';
 
 const SaveChartButton: FC<{ isExplorer?: boolean }> = ({ isExplorer }) => {
